feat: add multi-round simulation for long-term value estimation
when estimating the long term score of a move, several rounds up until the end of the game are simulated using average score and net gem profit values to get a more comprehensive view of its potential. so far this is experimental, since the bot has started to consider quite drastic plays, sacrificing 30+ points in some 6-10 gem situations in order to maintain a higher gem count throughout the entire game. i will be recomputingthe average values to account for this decision-making change